CID-3774: Addressing PR comments
